requirements.txt
Flask==3.0.0
pymysql==1.1.0
qrcode==7.4.2
Pillow==10.1.0
torch==2.1.0
transformers==4.35.0
pandas==2.1.3
openpyxl==3.1.2
reportlab==4.0.7
werkzeug==3.0.1
python-dotenv==1.0.0
